05-assignment/pyproject.toml
[project]
name = "05-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.10",
    "python-decouple>=3.8",
]
